fix: try to fix types from library
