+++ .storybook/main.ts
@@ -14,6 +14,10 @@ const config: StorybookConfig = {
     "@storybook/addon-essentials",
     "@storybook/addon-interactions",
   ],
+  typescript: {
+    reactDocgen: "react-docgen-typescript",
+    reactDocgenTypescriptOptions: {},
+  },
   docs: {
     defaultName: "Documentation",
   },
