+++ src/components/DateRangePicker/DateRangePicker.tsx
@@ -1,14 +1,8 @@
 import { FC } from "react";
-import { DefaultTheme } from "styled-components/dist/types";
 
 import Calendar from "@components/Calendar";
 import ThemeWrapper from "@components/ThemeWrapper";
 import ErrorBoundary from "@components/ErrorBoundary";
-import { CalendarProps } from "@components/Calendar/Calendar.types";
-
-interface DateRangePickerProps extends CalendarProps {
-  theme?: DefaultTheme;
-}
 
 export const DateRangePicker: FC<DateRangePickerProps> = ({
   theme,
