+++ src/decorators/withClearButton/index.tsx
@@ -1,7 +1,5 @@
 import { FC } from "react";
 
-import { CalendarProps } from "@components/Calendar/Calendar.types";
-
 const withClearButton = (Component: FC<CalendarProps>): FC<CalendarProps> => {
   const Wrappee = (props: CalendarProps) => {
     return <Component {...props} withClearButton={true} />;
