+++ src/types/index.d.ts
@@ -22,3 +22,30 @@ interface Todo {
 }
 
 type DateCallback = (date: Date) => void;
+
+interface CalendarProps {
+  dateSelected?: Date;
+  rangeStart?: Date;
+  rangeEnd?: Date;
+
+  onDateSelect?: (date: Date) => void;
+  onRangeStartSelect?: (date: Date) => void;
+  onRangeEndSelect?: (date: Date) => void;
+
+  withClearButton?: boolean;
+  weekStartsFromMonday?: boolean;
+  withTodos?: boolean;
+  withRangePicker?: boolean;
+
+  holidays?: Holiday[];
+  minYear?: number;
+  maxYear?: number;
+}
+
+interface DatePickerProps extends CalendarProps {
+  theme?: DefaultTheme;
+}
+
+interface DateRangePickerProps extends CalendarProps {
+  theme?: DefaultTheme;
+}
